Ice 3.7
Implements Slicer for Ice 3.7. This is largely:
* Pointers are now all std::shared_ptr
* libSlice isn't installed, so we build part of it into libslicer-compiler
* Library name change to use Ice++11

There are a few other things, but mostly just minor language/function
name compatibility changes.

The compiling of libSlice is a bit of a hack, currently comprising a git
submodule of Ice itself, and a few dirty defines on the command line,
but appears to be functionally satisfactory if a little unpleasant.